Fixed tool-tip font-size + other fixes
Tool-tips now display in the correct size on all breakpoints. Unnecessary color change onHover effect removed for angled headers. All text on the page when highlighted is white (as opposed to being blue before, making it invisible on the blue highlight background). Lastly, the width of the bootstrap container for high-res displays is now 80vw. This pushes the text in .Six, .Seven, & .Eight to the corners of the divs, where it belongs.