tool-tip fix + .Nine parallax effect
Tool-tips now display in the correct font size for all screen breakpoints. A new parallax scroll effect was added for .Nine. Unnecessary color change onHover effects were removed for rotated headers. Bootstrap container is now at 80vw for high-res screens, which pushes the text in .Six, .Seven, & .Eight to the edge of the div, where it belongs. Lastly, all text on the page if white when highlighted. Before, some of the text would be blue when highlighted, making it invisible against the blue highlight background.