Dynamic navbar + truly invisible arrows
Static arbitrary duration values for ScrollMagic navigation animations have been replaced. Now Performance, Exterior, Interior, & Specs headers stay highlighted for the exact number of pixels that their respective divs are in the viewport. The navbar also turns black for the exact number of pixels that Ten is in the viewport. When at the very bottom of the document, the navbar pulls now out of view allowing more of the background image to be viewed. Also, the arrows are now disabled after scrolling past the ScrollMagic trigger. Once each arrow starts to shrink & disappear, the arrows can no longer be clicked on & the tooltip will not show up when hovered over. Before this fix, #arrow-group could still be interacted with even after it disapppeared from view. This is no longer the case.