New BMW favicon + tweaks
The formerly missing site logo on the browser tab is now active. For consistency purposes, all headers in Specs.js are now underlined in blue, just like the headers in Footer.js. All checkmarks on the page are now also blue. Lastly, some description text was changed in Specs.js.